+++ bindings/python/Contacts/src/Contacts.cpp
@@ -10,8 +10,8 @@
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
 
-#include <iomanip>
 #include <chrono>
+#include <iomanip>
 
 #include <BipedalLocomotion/Contacts/Contact.h>
 #include <BipedalLocomotion/Contacts/ContactList.h>
@@ -86,6 +86,13 @@ void CreateContact(pybind11::module& module)
     py::class_<ContactWrench, ContactBase>(module, "ContactWrench")
         .def(py::init())
         .def_readwrite("wrench", &ContactWrench::wrench);
+    py::class_<DiscreteGeometryContact>(module, "DiscreteGeometryContact")
+        .def(py::init())
+        .def_readwrite("corners", &DiscreteGeometryContact::corners);
+    py::class_<Corner>(module, "Corner")
+        .def(py::init())
+        .def_readwrite("position", &Corner::position)
+        .def_readwrite("force", &Corner::force);
 }
 
 void CreateContactList(pybind11::module& module)
@@ -130,7 +137,8 @@ void CreateContactList(pybind11::module& module)
         .def("size", &ContactList::size)
         .def("__len__", &ContactList::size)
         .def("edit_contact", &ContactList::editContact, py::arg("element"), py::arg("new_contact"))
-        .def("get_present_contact",
+        .def(
+            "get_present_contact",
             [](const ContactList& l, const std::chrono::nanoseconds& time) -> PlannedContact {
                 return *l.getPresentContact(time);
             },
@@ -187,10 +195,29 @@ void CreateContactPhaseList(pybind11::module& module)
     using namespace BipedalLocomotion::Contacts;
     py::class_<ContactPhaseList>(module, "ContactPhaseList")
         .def(py::init())
+        .def("size", &ContactPhaseList::size)
+        .def("__getitem__", &ContactPhaseList::operator[])
+        .def(
+            "last_phase",
+            [](const ContactPhaseList& impl) -> const ContactPhase& { return *impl.lastPhase(); },
+            py::return_value_policy::reference_internal)
+        .def(
+            "first_phase",
+            [](const ContactPhaseList& impl) -> const ContactPhase& { return *impl.firstPhase(); },
+            py::return_value_policy::reference_internal)
         .def("set_lists",
              py::overload_cast<const ContactListMap&>(&ContactPhaseList::setLists),
              py::arg("contact_lists"))
-        .def("lists", &ContactPhaseList::lists);
+        .def("lists", &ContactPhaseList::lists)
+        .def(
+            "get_present_phase",
+            [](const ContactPhaseList& l, const std::chrono::nanoseconds& time)
+                -> const ContactPhase& { return *l.getPresentPhase(time); },
+            py::arg("time"),
+            py::return_value_policy::reference_internal)
+        .def("__iter__",
+             [](const ContactPhaseList& l) { return py::make_iterator(l.cbegin(), l.cend()); });
+    ;
 }
 
 void CreateContactListJsonParser(pybind11::module& module)